backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.3
geopandas==0.14.0
folium==0.14.0
plotly==5.18.0
numpy==1.26.2
shapely==2.0.2
python-dotenv==1.0.0
redis==5.0.1
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1
geoalchemy2==0.14.1
email-validator==2.0.0
psycopg2-binary==2.9.9
